Add: pagina1 page, add: pagina2 page, update: main
Add pagina1 page, this page will be the main page of the application, it has a floating action button that in the next updates will navigate to the page2.
Add pagina2 page, this page contains a scaffold with a green background color.
Update main, adding the pagina1 page to the home of the material app