Update: pagina1 page, update: pagina2 page, update: main
Update pagina1 page implementing the onPressed of the button, adding a navigator.push and creating a method called _crearRuta, this method will containg a PageRouteBuilder that returns a SlideTransition, with the purpose to animate the navigation between pages.
Update pagina2 adding some characteristics to the app bar and adding a center with the text Pagina2.
Update main adding the debug shoe banner in false.